Use arrays in mithril views
Mithril expects components at the same level of indentation to be in an
array. Sometimes it works as expected if I don't use an array, but it's
not consistent. I didn't want to use an array because I didn't want to
track open and closed brackets throughout the file. This pattern doesn't
require that.